requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
line-bot-sdk==3.11.0
schedule==1.2.0
pytz==2023.3